requirements.txt
# Core Dependencies
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5

# Data Processing
openpyxl==3.1.2
xlrd==2.0.1
scipy==1.11.2
statsmodels==0.14.0

# Machine Learning
imbalanced-learn==0.11.0
joblib==1.3.2

# Utilities
python-dotenv==1.0.0
pytest==7.4.0
pytest-cov==4.1.0

# API
requests==2.31.0
jsonschema==4.19.0

# Logging
loguru==0.7.2

# Time Series
prophet==1.1.4
nbformat==5.10.2

# Database (optional)
psycopg2-binary==2.9.7
redis==5.0.0